+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v1.4.0
+
+### Added
+
+ * Scheduled transaction support: `ScheduleCreateTransaction`, `ScheduleSignTransaction`, and `ScheduleSignTransaction`
+ * Implemented `TransactionList.getTransactions()` to support scheduling `ConsensusMessageSubmitTransaction`'s
+
+### Fixed
+
+* Decreased `ConsensusMessageSubmitTransaction` chunking size 4096->1024
+
 ## v1.1.4
 
 ### Added